+++ bin/replicante/src/components/mod.rs
@@ -1,3 +1,4 @@
+use failure::ResultExt;
 use prometheus::Registry;
 use slog::info;
 use slog::Logger;
@@ -6,12 +7,13 @@ use replicante_util_upkeep::Upkeep;
 
 use super::metrics::COMPONENTS_ENABLED;
 use super::Config;
+use super::Error;
+use super::ErrorKind;
 use super::Interfaces;
 use super::Result;
 
 mod core_api;
 mod discovery;
-mod events_indexer;
 mod grafana;
 mod update_checker;
 mod webui;
@@ -21,61 +23,86 @@ pub use self::discovery::Config as DiscoveryConfig;
 
 use self::core_api::CoreAPI;
 use self::discovery::DiscoveryComponent as Discovery;
-use self::events_indexer::EventsIndexer;
 use self::grafana::Grafana;
 use self::update_checker::UpdateChecker;
 use self::webui::WebUI;
 use self::workers::Workers;
 
-/// Helper function to keep `Components::new` simpler in the presence of optional components.
-fn init_component<C, F>(
-    components: &mut Vec<Box<dyn Component>>,
-    component: &str,
-    mode: &str,
-    enabled: bool,
-    logger: &Logger,
-    factory: F,
-) where
-    C: Component + 'static,
-    F: FnOnce() -> C,
-{
-    init_component_result(components, component, mode, enabled, logger, || {
-        Ok(factory())
-    })
-    .expect("component failed. Use init_component_result to propagate");
+/// Wrap a component defined ouside of this crate with the `Component` trait.
+///
+/// NOTE: this is a temporary solution to address two shortcomings:
+///   * Components should be defined externally but were not initially (they will be moved).
+///   * The codebase will go async in the future so interfaces will be redesigned then.
+macro_rules! impl_component {
+    ($name:ident, $real:ty) => {
+        struct $name($real);
+
+        impl Component for $name {
+            fn run(&mut self, upkeep: &mut Upkeep) -> Result<()> {
+                self.0
+                    .run(upkeep)
+                    .with_context(|_| ErrorKind::ComponentFailed(stringify!($name)))
+                    .map_err(Error::from)
+            }
+        }
+    };
+}
+
+impl_component!(ViewUpdater, replicore_component_viewupdater::ViewUpdater);
+impl ViewUpdater {
+    fn new(interfaces: &Interfaces) -> ViewUpdater {
+        let events = interfaces.streams.events.clone();
+        let logger = interfaces.logger.clone();
+        let store = interfaces.stores.view.clone();
+        let tracer = interfaces.tracing.tracer();
+        let component =
+            replicore_component_viewupdater::ViewUpdater::new(events, logger, store, tracer);
+        ViewUpdater(component)
+    }
 }
 
 /// Helper function to keep `Components::new` simpler in the presence of optional components.
-fn init_component_result<C, F>(
-    components: &mut Vec<Box<dyn Component>>,
-    component: &str,
-    mode: &str,
-    enabled: bool,
-    logger: &Logger,
-    factory: F,
-) -> Result<()>
-where
-    C: Component + 'static,
-    F: FnOnce() -> Result<C>,
-{
-    info!(
-        logger,
-        "Initialising component if enabled";
-        "component" => component,
-        "type" => mode,
-        "enabled" => enabled,
-    );
-    if enabled {
-        COMPONENTS_ENABLED
-            .with_label_values(&[component, mode])
-            .set(1.0);
-        components.push(Box::new(factory()?));
-    } else {
-        COMPONENTS_ENABLED
-            .with_label_values(&[component, mode])
-            .set(0.0);
-    }
-    Ok(())
+macro_rules! init_components {
+    {
+        let logger = $logger:expr;
+        $( component($name:literal, $mode:literal) {
+            let enabled = $enabled:expr;
+            $factory:expr
+        })+
+    } => {
+        {
+            let mut components: Vec<Box<dyn Component>> = Vec::new();
+            let logger = $logger;
+            $(
+                let enabled = $enabled;
+                if enabled {
+                    info!(
+                        logger,
+                        "Initialising component";
+                        "component" => $name,
+                        "type" => $mode,
+                        "enabled" => enabled,
+                    );
+                    COMPONENTS_ENABLED
+                        .with_label_values(&[$name, $mode])
+                        .set(1.0);
+                    components.push(Box::new($factory));
+                } else {
+                    info!(
+                        logger,
+                        "Skipped component";
+                        "component" => $name,
+                        "type" => $mode,
+                        "enabled" => enabled,
+                    );
+                    COMPONENTS_ENABLED
+                        .with_label_values(&[$name, $mode])
+                        .set(0.0);
+                }
+            )+
+            components
+        }
+    };
 }
 
 /// Generic replicante core component that does something.
@@ -100,70 +127,42 @@ pub struct Components {
 impl Components {
     /// Creates and configures components.
     pub fn new(config: &Config, logger: Logger, interfaces: &mut Interfaces) -> Result<Components> {
-        let mut components = Vec::new();
-        init_component(
-            &mut components,
-            "core_api",
-            "required",
-            config.components.core_api(),
-            &logger,
-            || CoreAPI::new(logger.clone(), interfaces),
-        );
-        init_component(
-            &mut components,
-            "discovery",
-            "required",
-            config.components.discovery(),
-            &logger,
-            || {
+        let components = init_components! {
+            let logger = &logger;
+            component("core_api", "required") {
+                let enabled = config.components.core_api();
+                CoreAPI::new(logger.clone(), interfaces)
+            }
+            component("discovery", "required") {
+                let enabled = config.components.discovery();
                 Discovery::new(
                     config.discovery.clone(),
                     config.events.snapshots.clone(),
                     logger.clone(),
                     interfaces,
                 )
-            },
-        );
-        init_component(
-            &mut components,
-            "events_indexer",
-            "required",
-            config.components.events_indexer(),
-            &logger,
-            || EventsIndexer::new(logger.clone(), interfaces),
-        );
-        init_component(
-            &mut components,
-            "grafana",
-            "optional",
-            config.components.grafana(),
-            &logger,
-            || Grafana::new(interfaces),
-        );
-        init_component(
-            &mut components,
-            "update_checker",
-            "optional",
-            config.components.update_checker(),
-            &logger,
-            || UpdateChecker::new(logger.clone()),
-        );
-        init_component(
-            &mut components,
-            "webui",
-            "optional",
-            config.components.webui(),
-            &logger,
-            || WebUI::new(interfaces),
-        );
-        init_component_result(
-            &mut components,
-            "workers",
-            "required",
-            config.components.workers(),
-            &logger,
-            || Workers::new(interfaces, logger.clone(), config.clone()),
-        )?;
+            }
+            component("grafana", "optional") {
+                let enabled = config.components.grafana();
+                Grafana::new(interfaces)
+            }
+            component("update_checker", "optional") {
+                let enabled = config.components.update_checker();
+                UpdateChecker::new(logger.clone())
+            }
+            component("viewupdater", "required") {
+                let enabled = config.components.viewupdater();
+                ViewUpdater::new(interfaces)
+            }
+            component("webui", "optional") {
+                let enabled = config.components.webui();
+                WebUI::new(interfaces)
+            }
+            component("workers", "required") {
+                let enabled = config.components.workers();
+                Workers::new(interfaces, logger.clone(), config.clone())?
+            }
+        };
         Ok(Components { components })
     }