+++ CHANGELOG.md
@@ -5,9 +5,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Populate view DB from the events stream.
+
 ### Fixed
 - Ensure `replictl` contexts store is flushed to disk before exiting.
 - WebUI `/cluster/{cluster_id}/action/{action_id}` handles `GET` requests instead of `POST`.
+
+### Removed
+- The `events_indexer` component was replaced by the `viewupdater` component.
 
 ## [0.6.0] - 2020-05-28
 ### Added